+++ image/service/keepalived/daemon.sh
@@ -1,2 +1,0 @@
-#!/bin/bash -e
-exec /usr/local/sbin/keepalived -f /etc/keepalived/keepalived.conf --dont-fork --log-console -D -d