+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## 0.2.0
-  - Upgrade baseimage: light-baseimage:0.2.0
+  - Upgrade baseimage: light-baseimage:0.2.1
   - Makefile with build no cache
 
 ## 0.1.9